+++ threedi_modelchecker/__init__.py
@@ -2,5 +2,5 @@ from .model_checks import *  # NOQA
 
 
 # fmt: off
-__version__ = '0.35.2.dev0'
+__version__ = '0.35.3.dev0'
 # fmt: on