+++ threedi_modelchecker/config.py
@@ -52,6 +52,7 @@ from .checks.raster import RasterRangeCheck
 from .checks.raster import RasterSquareCellsCheck
 from .checks.timeseries import TimeseriesIncreasingCheck
 from .checks.timeseries import TimeseriesRowCheck
+from .checks.timeseries import TimeseriesStartsAtZeroCheck
 from .checks.timeseries import TimeseriesTimestepCheck
 from .checks.timeseries import TimeseriesValueCheck
 from sqlalchemy import func
@@ -810,6 +811,12 @@ CHECKS += [
             "v2_exchange_line.the_geom is far (> 500 m) from its corresponding channel"
         ),
     ),
+    RangeCheck(
+        error_code=265,
+        column=models.ExchangeLine.exchange_level,
+        min_value=-9998.0,
+        max_value=8848.0,
+    ),
 ]
 
 ## 027x: Potential breaches
@@ -891,6 +898,19 @@ CHECKS += [
         level=CheckLevel.ERROR,
         column=models.PotentialBreach.the_geom,
         min_distance=TOLERANCE_M,
+    ),
+    RangeCheck(
+        error_code=276,
+        column=models.PotentialBreach.exchange_level,
+        min_value=-9998.0,
+        max_value=8848.0,
+    ),
+    RangeCheck(
+        error_code=277,
+        column=models.PotentialBreach.maximum_breach_depth,
+        min_value=0.0,
+        max_value=100.0,
+        left_inclusive=False,
     ),
 ]
 
@@ -2040,6 +2060,13 @@ CHECKS += [
         models.BoundaryConditions2D.timeseries,
         models.Lateral1d.timeseries,
         models.Lateral2D.timeseries,
+    ]
+]
+CHECKS += [
+    TimeseriesStartsAtZeroCheck(col, error_code=1204)
+    for col in [
+        models.BoundaryCondition1D.timeseries,
+        models.BoundaryConditions2D.timeseries,
     ]
 ]